Move to ConfigDB
a discussion with Mikee77 from the Sming project has led to him writing a ConfigDB library that allows to flexibly storing and retrieving config items (currently primarily with json in mind).
I will implement this as the standard config mechanism, replacing the global cfg structure (well, the app.cfg structure that is shared globally). While this will be a fair bit slower since access to config items will almost always require reading (or even writing) to flash (unless the current store is in cache), it should clean up the code a lot as the library can import and export json structures from and to streams and files directly and also reduce RAM usage as rather than holding a full json doc in memory to serialize/deserialize it plus holding the config struct in memory, this should work mostly on streams and much smaller blocks of json